+++ inhdelva/src/app/pages/rangoFactura/rangoFactura.component.ts
@@ -197,27 +197,40 @@ export class RangoFacturaComponent implements OnInit {
   }
 
   eliminar(data) {
-    this.rangoService.deleteRango(data.Id, { Estado: false })
-      .toPromise()
-      .then(
-        () => {
-          this.ShowNotification(
-            'success',
-            'Eliminado',
-            'El registro fue eliminado con éxito'
-          );
-          this.listOfDataRango = this.listOfDataRango.filter(x => x.Id !== data.Id);
-        },
-        (error) => {
+    swal({
+      title: "¿Está seguro de borrar el registro?",
+      // text: "Una vez eliminado el registro ",
+      icon: "warning",
+      buttons: true,
+      dangerMode: true,
+    })
+      .then((willDelete) => {
+        if (willDelete) {
+          this.rangoService.deleteRango(data.Id, { Estado: false })
+            .toPromise()
+            .then(
+              () => {
+                this.ShowNotification(
+                  'success',
+                  'Eliminado',
+                  'El registro fue eliminado con éxito'
+                );
+                this.listOfDataRango = this.listOfDataRango.filter(x => x.Id !== data.Id);
+              },
+              (error) => {
 
-          this.ShowNotification(
-            'error',
-            'No se pudo eliminar',
-            'El registro no pudo ser eleminado, por favor revise su conexión a internet o comuníquese con el proveedor.'
-          );
-          console.log(error);
+                this.ShowNotification(
+                  'error',
+                  'No se pudo eliminar',
+                  'El registro no pudo ser eleminado, por favor revise su conexión a internet o comuníquese con el proveedor.'
+                );
+                console.log(error);
+              }
+            );
+        } else {
+          swal("Su registro sigue activo");
         }
-      );
+      });
   }
 
 }
